reset counter when processing a new leaderboard
In the second period of any contest, for any given year, the board for
the current year would only have people in rank 1 (one needs 2 items to
be counted, and no-one can have more than 2 items).

The above means that after processing the current year board, prev would
be left equal to 1, and so the leader of the next board would show up
with rank '-' instead of 1 (because rank == prev).